# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "This is the API documentation for the Social API project",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/posts": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Get all posts",
                "description": "Retrieve a list of all posts.",
                "operationId": "getAllPosts",
                "responses": {
                    "200": {
                        "description": "A list of posts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "Post ID",
                                                "type": "integer"
                                            },
                                            "content": {
                                                "description": "Post content",
                                                "type": "string"
                                            },
                                            "image_url": {
                                                "description": "Image URL",
                                                "type": "string"
                                            },
                                            "profile": {
                                                "description": "Profile of the post author",
                                                "properties": {
                                                    "id": {
                                                        "description": "Profile ID",
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "description": "Profile name",
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "created_at": {
                                                "description": "Creation time",
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "Create a new post",
                "description": "Create a new post with content, an optional image, and link it to a profile.",
                "operationId": "createPost",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "profile_id",
                                    "content"
                                ],
                                "properties": {
                                    "profile_id": {
                                        "description": "Profile ID associated with the post",
                                        "type": "integer"
                                    },
                                    "content": {
                                        "description": "Post content",
                                        "type": "string"
                                    },
                                    "image_url": {
                                        "description": "Optional image file",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Post created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Success message",
                                            "type": "string"
                                        },
                                        "post": {
                                            "properties": {
                                                "id": {
                                                    "description": "Post ID",
                                                    "type": "integer"
                                                },
                                                "content": {
                                                    "description": "Post content",
                                                    "type": "string"
                                                },
                                                "image_url": {
                                                    "description": "Image URL",
                                                    "type": "string"
                                                },
                                                "profile": {
                                                    "description": "Profile of the post author",
                                                    "properties": {
                                                        "id": {
                                                            "description": "Profile ID",
                                                            "type": "integer"
                                                        },
                                                        "name": {
                                                            "description": "Profile name",
                                                            "type": "string"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/posts/{id}": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "Get a specific post",
                "description": "Retrieve a specific post by ID.",
                "operationId": "getPost",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "Post ID",
                                            "type": "integer"
                                        },
                                        "content": {
                                            "description": "Post content",
                                            "type": "string"
                                        },
                                        "image_url": {
                                            "description": "Image URL",
                                            "type": "string"
                                        },
                                        "profile": {
                                            "description": "Profile of the post author",
                                            "properties": {
                                                "id": {
                                                    "description": "Profile ID",
                                                    "type": "integer"
                                                },
                                                "name": {
                                                    "description": "Profile name",
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "created_at": {
                                            "description": "Creation time",
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Post"
                ],
                "summary": "Update an existing post",
                "description": "Update a post by ID, including content and optionally an image.",
                "operationId": "updatePost",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "content"
                                ],
                                "properties": {
                                    "content": {
                                        "description": "Post content",
                                        "type": "string"
                                    },
                                    "image_url": {
                                        "description": "Optional image URL",
                                        "type": "string",
                                        "format": "uri"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Post updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Success message",
                                            "type": "string"
                                        },
                                        "post": {
                                            "properties": {
                                                "id": {
                                                    "description": "Post ID",
                                                    "type": "integer"
                                                },
                                                "content": {
                                                    "description": "Post content",
                                                    "type": "string"
                                                },
                                                "image_url": {
                                                    "description": "Image URL",
                                                    "type": "string"
                                                },
                                                "profile": {
                                                    "description": "Profile of the post author",
                                                    "properties": {
                                                        "id": {
                                                            "description": "Profile ID",
                                                            "type": "integer"
                                                        },
                                                        "name": {
                                                            "description": "Profile name",
                                                            "type": "string"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Post"
                ],
                "summary": "Delete a post",
                "description": "Delete a post by ID.",
                "operationId": "deletePost",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Post ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Success message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Post not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your token here",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}